+++ internal/promapi/range.go
@@ -31,9 +31,10 @@ type rangeQuery struct {
 	ctx  context.Context
 	expr string
 	r    v1.Range
+	ttl  time.Duration
 }
 
-func (q rangeQuery) Run() (queryResult, int) {
+func (q rangeQuery) Run() queryResult {
 	log.Debug().
 		Str("uri", q.prom.safeURI).
 		Str("query", q.expr).
@@ -57,21 +58,18 @@ func (q rangeQuery) Run() (queryResult, int) {
 	resp, err := q.prom.doRequest(ctx, http.MethodPost, q.Endpoint(), args)
 	if err != nil {
 		qr.err = err
-		return qr, 1
+		return qr
 	}
 	defer resp.Body.Close()
 
 	if resp.StatusCode/100 != 2 {
 		qr.err = tryDecodingAPIError(resp)
-		return qr, 1
+		return qr
 	}
 
 	ranges, err := streamSampleStream(resp.Body, q.r.Step)
 	qr.value, qr.err = ranges, err
-	if ranges.Len() > 0 {
-		return qr, ranges.Len()
-	}
-	return qr, 1
+	return qr
 }
 
 func (q rangeQuery) Endpoint() string {
@@ -87,7 +85,7 @@ func (q rangeQuery) CacheKey() uint64 {
 }
 
 func (q rangeQuery) CacheTTL() time.Duration {
-	return 0
+	return q.ttl
 }
 
 type RangeQueryTimes interface {
@@ -140,6 +138,7 @@ func (p *Prometheus) RangeQuery(ctx context.Context, expr string, params RangeQu
 					End:   s.End,
 					Step:  step,
 				},
+				ttl: s.End.Sub(start) + time.Minute*10,
 			},
 		}