+++ internal/promapi/range_test.go
@@ -382,7 +382,7 @@ func TestRange(t *testing.T) {
 
 			fg := promapi.NewFailoverGroup("test", []*promapi.Prometheus{
 				promapi.NewPrometheus("test", srv.URL, nil, tc.timeout, 1, 100),
-			}, 1000, true, "up", nil, nil)
+			}, true, "up", nil, nil)
 			fg.StartWorkers()
 			defer fg.Close()